pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "UFS2ARCO"
authors = [
    { name = "Lauryn Eldridge", email = "leldridge@wvhtf.org" },
    { name = "Tim Smith", email = "tim.smith@noaa.gov" },
]
maintainers = [
    { name = "Lauryn Eldridge", email = "leldridge@wvhtf.org" },
    { name = "Tim Smith", email = "tim.smith@noaa.gov" },
    { name = "Ken Witt", email = "kwitt@wvhtf.org" },
]

description = "Initial data conversion pipeline"
readme = "README.rst"
requires-python = ">=3.9.17"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "PyYAML",
    "xarray",
    "numpy > 1.22",
    "packaging > 21.3",
    "pandas >1.4",
    "fsspec",
    "zarr",
    "cftime",
]

dynamic = ["version"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["UFS2ARCO*"]
exclude = ["tests*"]


[project.optional-dependencies]
dev = ["netcdf4", "h5netcdf"]
documentation = ["sphinx"]
test = ["black"]


[project.urls]
Documentation = "https://github.com/HighTechnologyFoundation/UFS2ARCO/tree/main#readme"
Source = "https://github.com/HighTechnologyFoundation/UFS2ARCO"
Tracker = "https://github.com/HighTechnologyFoundation/UFS2ARCO/issues"


[tool.setuptools_scm]
# For smarter version schemes and other configuration options,
# check out https://github.com/pypa/setuptools_scm
version_scheme = "no-guess-dev"

[tool.black]
line-length = 119

[tool.flake8]
max-line-length = 120
select = "F,E,W,B,B901,B902,B903"
exclude = [
    ".eggs",
    ".git",
    ".tox",
    "nssm",
    "obj",
    "out",
    "packages",
    "pywin32",
    "tests",
    "swagger_client",
    "build",
    "dist",
    ".rst",
]
ignore = [
    # "E722",
    "B001",
    "W503",
    "E203",
]
# "E722" - do not use bare except, specify exception instead
# "B001" 
# "W503" - line break before binary operator
# "E203" - whitespace before ‘,’, ‘;’, or ‘:’

max-complexity = 10


[tool.bumpver]
current_version = "0.0.2"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "Bump version {old_version} -> {new_version}"
commit = true
tag = true
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = ['^current_version = "{version}"']
"src/UFS2ARCO/__init__.py" = ['^__version__ = "{version}"']
